Separate Factory from VM.
