+++ examples/fibo.js
@@ -1,6 +1,6 @@
-function fibo(x) {
+function fibo (x) {
   if (x < 2) return 1
   return fibo(x - 1) + fibo(x - 2)
 }
 
-console.log(fibo(22))
+console.log(fibo(30))
