+++ src/vm/mod.rs
@@ -4,4 +4,5 @@ pub mod codegen;
 pub mod constant;
 pub mod error;
 pub mod exec_context;
+pub mod factory;
 pub mod vm;
