+++ test/fibo.js
@@ -1,8 +1,8 @@
 let assert = require('assert').deepStrictEqual
 
-function fibo(x) {
+function fibo (x) {
   if (x < 2) return 1
   return fibo(x - 1) + fibo(x - 2)
 }
 
-assert(fibo(20), 10946)
+assert(fibo(10), 89)
